+++ lib/database_consistency/processors/validators_processor.rb
@@ -5,7 +5,6 @@ module DatabaseConsistency
     # The class to process validators
     class ValidatorsProcessor < BaseProcessor
       CHECKERS = [
-        Checkers::BelongsToPresenceChecker,
         Checkers::MissingUniqueIndexChecker
       ].freeze